GPLIM-6607 changing decimal initialization on Big Decimals to do so in strings

+++ mercury/src/test/java/org/broadinstitute/gpinformatics/athena/boundary/billing/BillingCreditDbFreeTest.java
@@ -104,7 +104,7 @@ public class BillingCreditDbFreeTest {
     private BillingEjb billingEjb;
 
     private final BigDecimal qtyPositiveTwo = BigDecimal.valueOf(2);
-    private final BigDecimal qtyNegativeTwo = BigDecimal.valueOf(qtyPositiveTwo.longValue()).negate();
+    private final BigDecimal qtyNegativeTwo = qtyPositiveTwo.negate();
 
     @BeforeMethod
     public void setUp()
